src/database/user_model.py
from sqlalchemy import Column, Integer, String, DateTime
from datetime import datetime
from src.database.db_config import Base

class User(Base):
    """User tabel for the database."""

    __tablename__ = "users"

    id = Column(Integer, primary_key=True, index=True)
    email = Column(String(255), unique=True,nullable=False, index=True)
    hashed_password = Column(String(255), nullable=False)
    full_name = Column(String(255), nullable=True)
    created_at = Column(DateTime, default=datetime.now)
    updated_at = Column(DateTime, default=datetime.now, onupdate=datetime.now)
